+++ src/upgrades/mongos_upgrade.py
@@ -4,29 +4,26 @@
 """Manager for handling Mongos in-place upgrades."""
 
 import logging
-import secrets
-import string
-from typing import Optional, Tuple
-
+from typing import Optional
+from ops import BlockedStatus
 from ops.charm import ActionEvent, CharmBase
-from ops.framework import Object, EventBase, EventSource
-from ops.model import ActiveStatus, BlockedStatus
-from tenacity import retry, stop_after_attempt, wait_fixed
-
-from charms.mongodb.v1.mongos import (
-    MongosConnection,
+from ops.framework import EventBase, EventSource
+from ops.model import ActiveStatus
+from overrides import override
+from charms.mongodb.v0.upgrade_helpers import (
+    ROLLBACK_INSTRUCTIONS,
+    GenericMongoDBUpgrade,
+    PEER_RELATION_ENDPOINT_NAME,
+    PeerRelationNotReady,
+    UnitState,
+    PrecheckFailed,
 )
 
 from config import Config
-from upgrades import machine_upgrade, upgrade
+from upgrades import machine_upgrade
 
 logger = logging.getLogger(__name__)
 
-
-WRITE_KEY = "write_value"
-ROLLBACK_INSTRUCTIONS = "To rollback, `juju refresh` to the previous revision"
-
-SHARD_NAME_INDEX = "_id"
 
 # BEGIN: Exceptions
 
@@ -45,26 +42,40 @@ class _PostUpgradeCheckMongos(EventBase):
         super().__init__(handle)
 
 
-class MongosUpgrade(Object):
+class MongosUpgrade(GenericMongoDBUpgrade):
     """Handlers for upgrade events."""
 
     post_upgrade_event = EventSource(_PostUpgradeCheckMongos)
 
     def __init__(self, charm: CharmBase):
         self.charm = charm
-        super().__init__(charm, upgrade.PEER_RELATION_ENDPOINT_NAME)
+        super().__init__(charm, PEER_RELATION_ENDPOINT_NAME)
+
+    @override
+    def _observe_events(self, charm: CharmBase) -> None:
         self.framework.observe(
-            charm.on[upgrade.PEER_RELATION_ENDPOINT_NAME].relation_created,
+            charm.on[PEER_RELATION_ENDPOINT_NAME].relation_created,
             self._on_upgrade_peer_relation_created,
         )
         self.framework.observe(
-            charm.on[upgrade.PEER_RELATION_ENDPOINT_NAME].relation_changed,
+            charm.on[PEER_RELATION_ENDPOINT_NAME].relation_changed,
             self._reconcile_upgrade,
         )
         self.framework.observe(charm.on.upgrade_charm, self._on_upgrade_charm)
 
         self.framework.observe(charm.on["force-upgrade"].action, self._on_force_upgrade_action)
         self.framework.observe(self.post_upgrade_event, self.run_post_upgrade_check)
+
+    # BEGIN: properties
+    @property
+    @override
+    def _upgrade(self) -> Optional[machine_upgrade.Upgrade]:
+        try:
+            return machine_upgrade.Upgrade(self.charm)
+        except PeerRelationNotReady:
+            return None
+
+    # END: properties
 
     # BEGIN: Event handlers
     def _on_upgrade_peer_relation_created(self, _) -> None:
@@ -89,10 +100,10 @@ class MongosUpgrade(Object):
         if not self._upgrade.is_compatible:
             self._set_upgrade_status()
             return
-        if self._upgrade.unit_state is upgrade.UnitState.OUTDATED:
+        if self._upgrade.unit_state is UnitState.OUTDATED:
             try:
                 authorized = self._upgrade.authorized
-            except upgrade.PrecheckFailed as exception:
+            except PrecheckFailed as exception:
                 self._set_upgrade_status()
                 self.charm.status.set_and_share_status(exception.status)
                 logger.debug(f"Set unit status to {self.unit.status}")
@@ -136,7 +147,7 @@ class MongosUpgrade(Object):
         # The mongos service cannot be considered ready until it has a config-server. Therefore
         # it is not necessary to do any sophisticated checks.
         if not self.charm.mongos_intialised:
-            self._upgrade.unit_state = upgrade.UnitState.HEALTHY
+            self._upgrade.unit_state = UnitState.HEALTHY
             return
 
         self.run_post_upgrade_checks(event)
@@ -144,6 +155,25 @@ class MongosUpgrade(Object):
     # END: Event handlers
 
     # BEGIN: Helpers
+    def _set_upgrade_status(self):
+        # In the future if we decide to support app statuses, we will need to handle this
+        # differently. Specifically ensuring that upgrade status for apps status has the lowest
+        # priority
+        if self.charm.unit.is_leader():
+            self.charm.app.status = self._upgrade.app_status or ActiveStatus()
+
+        # Set/clear upgrade unit status if no other unit status - upgrade status for units should
+        # have the lowest priority.
+        if isinstance(self.charm.unit.status, ActiveStatus) or (
+            isinstance(self.charm.unit.status, BlockedStatus)
+            and self.charm.unit.status.message.startswith(
+                "Rollback with `juju refresh`. Pre-upgrade check failed:"
+            )
+        ):
+            self.charm.status.set_and_share_status(
+                self._upgrade.get_unit_juju_status() or ActiveStatus()
+            )
+
     def run_post_upgrade_checks(self, event) -> None:
         """Runs post-upgrade checks for after a shard/config-server/replset/cluster upgrade."""
         logger.debug("-----\nchecking mongos running\n----")
@@ -164,16 +194,6 @@ class MongosUpgrade(Object):
             self.charm.status.set_and_share_status(ActiveStatus())
 
         logger.debug("upgrade of unit succeeded.")
-        self._upgrade.unit_state = upgrade.UnitState.HEALTHY
+        self._upgrade.unit_state = UnitState.HEALTHY
 
     # END: helpers
-
-    # BEGIN: properties
-    @property
-    def _upgrade(self) -> Optional[machine_upgrade.Upgrade]:
-        try:
-            return machine_upgrade.Upgrade(self.charm)
-        except upgrade.PeerRelationNotReady:
-            pass
-
-    # END: properties
